feat: basic init commit
This is just to get something on the page.
All content was generated by chatgpt and I haven't had time to look over it and proof it.
From a quick glance I believe there can be improvement by adding links to the home page on the children pages and there is probably a better way to do the headers and footers
EVERYTHING IS SUBJECT TO CHANGE!!! NOTHING HERE WAS MENT TO BE PERMANENT